pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dotenv-tools"
version = "0.0.1"
description = "A comprehensive CLI tool to manage environment variables in .env files with complete syntax support"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "LousyBook01", email = "lousybook94@gmail.com"}
]
maintainers = [
    {name = "LousyBook01", email = "lousybook94@gmail.com"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Shells",
    "Topic :: System :: Systems Administration",
]
keywords = ["dotenv", "environment", "cli", "env", "environment-variables"]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.scripts]
load-dotenv = "dotenv_tools.cli:load_dotenv"
unload-dotenv = "dotenv_tools.cli:unload_dotenv"
set-dotenv = "dotenv_tools.cli:set_dotenv"
export-dotenv = "dotenv_tools.cli:export_dotenv_cmd"
generate-template = "dotenv_tools.cli:generate_template_cmd"
compare-env = "dotenv_tools.cli:compare_env_cmd"
shell-completion = "dotenv_tools.cli:shell_completion_cmd"

[project.urls]
Homepage = "https://github.com/LousyBook94/dotenv-tools"
Repository = "https://github.com/LousyBook94/dotenv-tools"
Issues = "https://github.com/LousyBook94/dotenv-tools/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/dotenv_tools --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
